Reduce duplicated mocks for the global variable 'CourseList'.
